requirements.txt
pdfminer.six
PyPDF2
pillow
pytesseract
streamlit